Adjust get_completion_badge usages to handle exceptions.

+++ lms/djangoapps/certificates/management/commands/regenerate_user.py
@@ -111,11 +111,13 @@ class Command(BaseCommand):
                 course_id
             )
 
-            badge_class = get_completion_badge(course_id, student)
-            badge = badge_class.get_for_user(student)
-            if badge:
-                badge.delete()
-                LOGGER.info(u"Cleared badge for student %s.", student.id)
+            if course.issue_badges:
+                badge_class = get_completion_badge(course_id, student)
+                badge = badge_class.get_for_user(student)
+
+                if badge:
+                    badge.delete()
+                    LOGGER.info(u"Cleared badge for student %s.", student.id)
 
             # Add the certificate request to the queue
             ret = regenerate_user_certificates(
